requirements.txt
django==2.1
djangorestframework==3.8.2
django-rest-auth==0.9.3
django-allauth==0.37.1
